# RFC: auto-fixes from ruff 0.6 (io.fits)