@pllim 2024-08-21
Hmm benchmark failures seem unrelated? 🤷‍♀️ 

```
| Change   | Before [81432542]    | After [90d70649]    |   Ratio | Benchmark (Parameter)          |
|----------|----------------------|---------------------|---------|--------------------------------|
| +        | 35.2±0.07μs          | 123±70μs            |    3.49 | units.time_quantity_creation   |
| +        | 93.2±0.5μs           | 223±100μs           |    2.39 | units.time_quantity_init_array |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```