@mhvk 2024-08-27
Here, I really wondered about the code, but since this PR is just about refactoring to use max/min, that should not stop it.